move the `syntax` imports to the xsp package object
pretty much so you can import xsp._ and be done with it
