+++ core2/src/main/scala/io/dylemma/xsp/package.scala
@@ -1,0 +1,11 @@
+package io.dylemma
+
+import io.dylemma.xsp.syntax._
+
+package object xsp
+	extends TransformerSyntax
+	with ContextMatcherSyntax {
+
+	object TransformerSyntax extends TransformerSyntax
+	object ContextMatcherSyntax extends ContextMatcherSyntax
+}
