--- core2/src/main/scala/io/dylemma/xsp/syntax.scala
@@ -1,5 +1,0 @@
-package io.dylemma.xsp
-
-object syntax
-	extends ContextMatcherSyntax
-	with TransformerSyntax
