--- examples/src/main/resources/example-comments2.xml
@@ -1,22 +1,0 @@
-<comments>
-	<comment date="2014-07-20">
-		<user name="alice" id="98qja34j3"/>
-		<stats upvote-count="123" downvote-count="20"/>
-		<body>Parsing XML sure is a pain...</body>
-	</comment>
-	<comment date="2014-07-21">
-		<user name="bob" id="3adfngfl"/>
-		<stats upvote-count="10" downvote-count="2"/>
-		<body>Yeah, I use SAX, but it's still pretty tedious</body>
-	</comment>
-    <comment date="2014-07-21">
-        <user name="alice" id="98qja34j3"/>
-        <stats upvote-count="15" downvote-count="3"/>
-        <body>I tried StAX. It seems nicer, but still fairly repetitive</body>
-    </comment>
-	<comment date="2015-10-11">
-		<user name="dylemma" id="038oqfdje"/>
-		<stats upvote-count="23841" downvote-count="4"/>
-		<body>I made a new library to make XML parsing less awful! Try xml-stream, it'll change your life.</body>
-	</comment>
-</comments>
