--- examples/src/main/resources/example-complex.xml
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<findings>
-    <finding severity="high" status="new">
-        <location path="somewhere/neat/but/bad.java">
-            <line start="3"/>
-            <column start="10" end="15"/>
-        </location>
-        <comments>
-            <comment user="dylan">
-                This is a *very* good comment!
-            </comment>
-        </comments>
-    </finding>
-
-    <finding severity="medium" status="unresolved">
-        <location path="somewhere/else.txt">
-            <line start="6" end="10"/>
-        </location>
-        <comments>
-            <comment user="baf">
-                Hi there
-            </comment>
-            <comment user="dylan">
-                Oh hello
-            </comment>
-        </comments>
-    </finding>
-
-    <finding severity="low" status="false-positive">
-        <location path="beyond/the/rainbow.mp4"/>
-    </finding>
-</findings>
