+++ examples/src/main/scala/io/dylemma/xml/example/Example2_Contexts.scala
@@ -1,7 +1,6 @@
 package io.dylemma.xml.example
 
 import io.dylemma.xsp._
-import io.dylemma.xsp.syntax._
 
 object Example2_Contexts extends App {
