+++ examples/src/main/scala/io/dylemma/xml/example/Example3_EntryList.scala
@@ -1,7 +1,6 @@
 package io.dylemma.xml.example
 
 import io.dylemma.xsp._
-import io.dylemma.xsp.syntax._
 
 /**
  * Created by dylan on 11/24/2015.
